# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Time Off Service API
  version: 0.1.0
  description: |
    MyCompany time-off microservice. JSON:API v1.1 over `application/vnd.api+json`.

    **Authentication:** Bearer JWT (HS256). Mint test tokens with `npm run tools:jwt`.

    **Correlation:** Send optional `X-Correlation-Id`; echoed in responses and error `meta`.

    Documents **implemented** Phase 1 and Phase 2 endpoints. See [spec.md](spec.md) for the full roadmap.
  contact:
    name: Time Off Service
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/sync/time-off:
    post:
      tags: [Sync]
      summary: Trigger HCM sync
      description: Runs the full HCM batch sync pipeline (employees, leave types, policies, balances).
      operationId: triggerSync
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Sync started/completed
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/SyncRunDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Sync already in progress
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/JsonApiErrorDocument'
  /api/v1/sync/status:
    get:
      tags: [Sync]
      summary: Get sync status
      description: Last sync metadata, staleness, and counts.
      operationId: getSyncStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Current sync status
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/SyncStatusDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/reports/leave-usage:
    get:
      tags: [Reports]
      summary: Leave usage report
      operationId: getLeaveUsageReport
      security:
        - bearerAuth: []
      parameters:
        - name: filter[startDate]
          in: query
          schema:
            type: string
            format: date
        - name: filter[endDate]
          in: query
          schema:
            type: string
            format: date
        - name: filter[department]
          in: query
          schema:
            type: string
        - name: filter[employeeId]
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Usage summary with meta.summary aggregates
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/reports/team-calendar:
    get:
      tags: [Reports]
      summary: Team calendar report
      operationId: getTeamCalendarReport
      security:
        - bearerAuth: []
      parameters:
        - name: filter[startDate]
          in: query
          schema:
            type: string
            format: date
        - name: filter[endDate]
          in: query
          schema:
            type: string
            format: date
        - name: filter[employeeId]
          in: query
          schema:
            type: string
            format: uuid
        - name: filter[teamId]
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Team leave calendar entries
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/reports/audit:
    get:
      tags: [Reports]
      summary: Audit export
      operationId: getAuditReport
      security:
        - bearerAuth: []
      parameters:
        - name: filter[startDate]
          in: query
          schema:
            type: string
            format: date
        - name: filter[endDate]
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Paginated audit log export (no email in rows)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. Claims: `sub`, `roles[]`, optional `employeeId`.
        Roles: `employee`, `manager`, `hr_admin`, `system_admin`, `integration_client`.
  parameters:
...
    PageNumber:
      name: page[number]
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page[size]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
      description: Optional trace ID; generated if omitted
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Optional idempotency key for safe POST retries (24h TTL)
  headers:
    CorrelationId:
      description: Request correlation ID
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/vnd.api+json:
          schema:
            $ref: '#/components/schemas/JsonApiErrorDocument'
    Forbidden:
      description: Insufficient permissions
      content:
        application/vnd.api+json:
          schema:
            $ref: '#/components/schemas/JsonApiErrorDocument'
...
  schemas:
...
    JsonApiVersion:
      type: object
      required: [version]
      properties:
        version:
          type: string
          example: '1.1'
    JsonApiError:
      type: object
      properties:
        status:
          type: string
        code:
          type: string
        title:
          type: string
        detail:
          type: string
        source:
          type: object
          properties:
            pointer:
              type: string
    JsonApiErrorDocument:
      type: object
      required: [jsonapi, errors]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/JsonApiError'
        meta:
          type: object
          properties:
            correlationId:
              type: string
...
    SyncRunAttributes:
      type: object
      properties:
        syncType:
          type: string
        syncRunId:
          type: string
          format: uuid
        status:
          type: string
          enum: [SUCCESS, PARTIAL, FAILED, IN_PROGRESS]
        correlationId:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        employeeCount:
          type: integer
        balanceCount:
          type: integer
        adjustmentCount:
          type: integer
        errorDetails:
          type: object
        createdAt:
          type: string
          format: date-time
    SyncRunResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [sync-runs]
        id:
          type: string
          format: uuid
        attributes:
          $ref: '#/components/schemas/SyncRunAttributes'
    SyncRunDocument:
      type: object
      required: [jsonapi, data]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        data:
          $ref: '#/components/schemas/SyncRunResource'
        meta:
          type: object
          properties:
            correlationId:
              type: string
...
    SyncStatusAttributes:
      type: object
      properties:
        syncSource:
          type: string
        lastSyncStartedAt:
          type: string
          format: date-time
        lastSyncCompletedAt:
          type: string
          format: date-time
        lastSyncStatus:
          type: string
        employeeCount:
          type: integer
        balanceCount:
          type: integer
        stalenessSeconds:
          type: integer
    SyncStatusResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [sync-status]
        id:
          type: string
          example: current
        attributes:
          $ref: '#/components/schemas/SyncStatusAttributes'
    SyncStatusDocument:
      type: object
      required: [jsonapi, data]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        data:
          $ref: '#/components/schemas/SyncStatusResource'